Deleted PhotoFrame & renamed PhotoFrameInGrid to PhotoFrame, updated conditional rendering in PhotoGrid & styling in NoResults
