+++ src/components/UIComponents/NoResults/NoResults.scss
@@ -1,8 +1,8 @@
 #no-matches-wrapper {
-    height: 40vh;
+    height: 33vh;
     display: flex;
     justify-content: center;
-    align-items: center;
+    align-items: flex-end;
     grid-column: 1 / -1;
 
     p {
