+++ src/index.tsx
@@ -10,9 +10,9 @@ const root = ReactDOM.createRoot(
 );
 
 root.render(
-    // <React.StrictMode>
+    <React.StrictMode>
         <Provider store={store}>
             <App />
         </Provider>
-    /* </React.StrictMode>  */
+    </React.StrictMode> 
 );
